Version 1.1:
NEW:
-Faster NMAP scan.
-Batch mode, run the scan without the user's input.
-Help menu redesigned.
-Added an option to pass a custom wordlist to Gobuster (In case the default one isn't found)

Fixes:
-Fixed detection of SMB, FTP and SSH open ports.
-Fixed "blank" in Tun0 IP when not connected to a VPN.
-Minor bugs fixed as well...

+++ YAAS.sh
@@ -24,7 +24,7 @@ yellow="\e[1;33m" # Yellow
 reset='\e[0m' # Text Reset
 
 #Generals
-version="1.0"
+version="1.1"
 
 if [[ user_folder == "" ]]; then
 	user_folder= $(pwd)
@@ -116,12 +116,14 @@ n_flag=''
 i_flag=''
 f_flag=false
 h_flag=false
+b_flag=false
 
-while getopts 'n:i:fh' flag; do
+while getopts 'n:i:fhb' flag; do
   case "${flag}" in
     n) n_flag="${OPTARG}" ;;
 	i) i_flag="${OPTARG}" ;;
     f) f_flag='true' ;;
+	b) b_flag='true' ;;
     h) h_flag= print_help 
 		exit 1 ;;
   esac
@@ -146,32 +148,32 @@ else
 	read -p "[+] IP : " ip
 fi
 
+hugo_created=${red}NO${reset}
+
 ##Add an specific hostname for the given IP
-read -p "[+] You want to add an specific hostname for the IP ? [y/N] : " add_host
-if [[ $add_host == "yes" || $add_host == "y" || $add_host == "YES" || $add_host == "Y" ]]; then
-	read -p " ╰─ Please write it here : " name_host
-	write_hosts=`echo "${ip} ${name_host}" | sudo tee -a /etc/hosts`
-fi
-
-#Ask for Hugo page
-read -p "[+] Create Page [y/N]: " hugo_page
-if [[ $name_folder != "" && $hugo_page == "yes" || $hugo_page == "y" || $hugo_page == "YES" || $hugo_page == "Y" ]]; then
-	if [ -f "$hugo_dir" ]; then
-		command cd "${hugo_dir}"
-	else
-		until [ -d "$hugo_input" ]; do
-			box_out "PATH FOR HUGO NOT SET"
-			read -p "[+] Hugo Path [$HOME] : " hugo_input
-		done
-		hugo_dir=$hugo_input
-		command cd "${hugo_dir}"
+if [[ $b_flag = false ]]; then
+	read -p "[+] You want to add an specific hostname for the IP ? [y/N] : " add_host
+	if [[ $add_host == "yes" || $add_host == "y" || $add_host == "YES" || $add_host == "Y" ]]; then
+		read -p " ╰─ Please write it here : " name_host
+		write_hosts=`echo "${ip} ${name_host}" | sudo tee -a /etc/hosts`
 	fi
-	command hugo new "posts/$name_folder/index.md"
-	hugo_created=${green}YES${reset}
-else
-	hugo_created=${red}NO${reset}
+	#Ask for Hugo page
+	read -p "[+] Create Page [y/N]: " hugo_page
+	if [[ $name_folder != "" && $hugo_page == "yes" || $hugo_page == "y" || $hugo_page == "YES" || $hugo_page == "Y" ]]; then
+		if [ -f "$hugo_dir" ]; then
+			command cd "${hugo_dir}"
+		else
+			until [ -d "$hugo_input" ]; do
+				box_out "PATH FOR HUGO NOT SET"
+				read -p "[+] Hugo Path [$HOME] : " hugo_input
+			done
+			hugo_dir=$hugo_input
+			command cd "${hugo_dir}"
+		fi
+		command hugo new "posts/$name_folder/index.md"
+		hugo_created=${green}YES${reset}
+	fi
 fi
-
 command cd "${user_folder}"
 
 #Create the folder or use the created one
@@ -182,7 +184,6 @@ if [[ $name_folder != "" ]]; then
 		command cd "${user_folder}${name_folder}"
 	fi
 fi
-
 #Check if we have all the necessary tools
 tool_checker
 
@@ -193,20 +194,26 @@ if [[ $ip != "" ]]; then
 	if [[ $(ping -c 2 $ip 2> /dev/null) ]]; then
 		echo -e "${lgreen}Everything is UP and running${reset}"
 		ip_tun0=`hostname -I | awk '{print $2}'`
+		if [[ $ip_tun0 == "" ]]; then
+			ip_tun0="N/A"
+		fi
 	else
-		echo -e "${red}Host is not accepting ping (Windows?) or unknown hostname${reset}"
-		read -p "[+] Continue? [Y/n] : " continue_ping
-		if [[ $continue_ping == "no" || $continue_ping == "n" || $continue_ping == "NO" || $continue_ping == "N" ]]; then
-			echo ""
-			echo -e "${red}Exiting now...${reset}"
-			exit 1;
+		if [[ "$b_flag" = false ]]; then
+			echo -e "${red}Host is not accepting ping (Windows?) or unknown hostname${reset}"
+			read -p "[+] Continue? [Y/n] : " continue_ping
+			if [[ $continue_ping == "no" || $continue_ping == "n" || $continue_ping == "NO" || $continue_ping == "N" ]]; then
+				echo ""
+				echo -e "${red}Exiting now...${reset}"
+				exit 1;
+			fi
 		fi
 	fi
 	box_out "Launching NMAP Scan..."
 	if [[ "$f_flag" = true ]]; then
-		command nmap $ip -o output_nmap
+		command sudo nmap $ip -p- -sS --min-rate 5000 -n -Pn -o nmap_Scan
+		open_ports=`cat nmap_Scan | grep open | awk -F/ '{print $1}' ORS=',' | rev | cut -c 2- | rev`
 	else
-		first_nmap=`nmap $ip -o output_nmap`
+		first_nmap=`sudo nmap $ip -p- -sS --min-rate 5000 -n -Pn -o output_nmap`
 		open_ports=`cat output_nmap | grep open | awk -F/ '{print $1}' ORS=',' | rev | cut -c 2- | rev`
 		sudo nmap -sS -sV -sC $ip -p $open_ports -o nmap_Scan
 	fi
@@ -219,7 +226,7 @@ ftp=false
 ftp_anon=false
 ssh=false
 
-for line in $(cat output_nmap | grep open)
+for line in $(cat nmap_Scan | grep open)
 do
 	if [[ $line =~ (http|web|server) ]]; then
 		web_server=true
@@ -235,81 +242,80 @@ do
 	fi
 done
 
+if [[ $b_flag = false ]]; then
+	#IF we have a web server, ask for the port and launch a common gobuster
+	if [[ "$web_server" = true ]]; then
 
-#IF we have a web server, ask for the port and launch a common gobuster
-if [[ "$web_server" = true ]]; then
-
-	box_out "WEB SERVER DETECTED"
-	read -p "[+] Scan Web Server? [y/N] : " scan_web
-	if [[ $scan_web == "yes" || $scan_web == "y" || $scan_web == "Y" || $scan_web == "YES" ]]; then
-		read -p "[+] Port? [80] : " port_web
-		if [[ $port_web == "" ]]; then
-			port_web=80
-		fi
-		if [[ $gobuster_dic == "" ]]; then
-			if [ -f "/usr/share/wordlists/dirb/common.txt" ]; then
-				gobuster_dic="/usr/share/wordlists/dirb/common.txt"
-			else
-				until [ -f "$gobuster_input" ]; do
-					box_out "DICTIONARY FOR GOBUSTER NOT FOUND"
-					read -p "[+] Dictionary Path [/usr/share/wordlists/dirb/common.txt] : " gobuster_input
-				done
+		box_out "WEB SERVER DETECTED"
+		read -p "[+] Scan Web Server? [y/N] : " scan_web
+		if [[ $scan_web == "yes" || $scan_web == "y" || $scan_web == "Y" || $scan_web == "YES" ]]; then
+			read -p "[+] Port? [80] : " port_web
+			if [[ $port_web == "" ]]; then
+				port_web=80
 			fi
-		else
-			if [ -f "$gobuster_dic" ]; then
-				gobuster_dic=$gobuster_dic
-			else
-				until [ -f "$gobuster_input" ]; do
-					box_out "DICTIONARY FOR GOBUSTER NOT FOUND"
-					read -p "[+] Dictionary Path [/usr/share/wordlists/dirb/common.txt] : " gobuster_input
-				done
-				gobuster_dic=$gobuster_input
-			fi
-		fi
-
-		gobuster dir -w ${gobuster_dic} --threads 50 -x txt,old,bak,zip,php,html,txt,bak,old,pdf -u "http://$ip:$port_web" 2> /dev/null | tee Gobuster_Scan.txt
-		cat Gobuster_Scan.txt | grep '301\|200' |  awk -F/ '{print $2}'| sed -e 's/\s.*$//' > endpoints.txt
-
-		for endpoint in $(cat endpoints.txt)
-		do 
-			result_curl=`curl -s http://$ip:$port_web/$endpoint > searchingfor_WP.txt`
-			if [[ $(cat searchingfor_WP.txt | grep Moved) ]]; then
-				result_curl=`curl -s http://$ip:$port_web/$endpoint/ > searchingfor_WP.txt`
-			fi
-			if [[ $(cat searchingfor_WP.txt | grep wordpress) ]]; then
-				endpoint_wp=${endpoint}
-				box_out "We detected a potential WordPress page"
-				echo "[!] Endpoint Name : " $endpoint_wp
-				read -p "[+] Scan WordPress? [y/N] : " scan_wp
-				if [[ $scan_wp == "yes" || $scan_wp == "y" || $scan_wp == "Y" || $scan_wp == "YES" ]]; then
-					wpscan --no-update -o WP_SCAN.json -f json -e vp,vt,cb,u --url "http://$ip:$port/$endpoint_wp"
+			if [[ $gobuster_dic == "" ]]; then
+				if [ -f "/usr/share/wordlists/dirb/common.txt" ]; then
+					gobuster_dic="/usr/share/wordlists/dirb/common.txt"
+				else
+					until [ -f "$gobuster_input" ]; do
+						box_out "DICTIONARY FOR GOBUSTER NOT FOUND"
+						read -p "[+] Dictionary Path [/usr/share/wordlists/dirb/common.txt] : " gobuster_input
+					done
 				fi
-				break
+			else
+				if [ -f "$gobuster_dic" ]; then
+					gobuster_dic=$gobuster_dic
+				else
+					until [ -f "$gobuster_input" ]; do
+						box_out "DICTIONARY FOR GOBUSTER NOT FOUND"
+						read -p "[+] Dictionary Path [/usr/share/wordlists/dirb/common.txt] : " gobuster_input
+					done
+					gobuster_dic=$gobuster_input
+				fi
 			fi
-		done
+
+			gobuster dir -w ${gobuster_dic} --threads 50 -x txt,old,bak,zip,php,html,txt,bak,old,pdf -u "http://$ip:$port_web" 2> /dev/null | tee Gobuster_Scan.txt
+			cat Gobuster_Scan.txt | grep '301\|200' |  awk -F/ '{print $2}'| sed -e 's/\s.*$//' > endpoints.txt
+
+			for endpoint in $(cat endpoints.txt)
+			do 
+				result_curl=`curl -s http://$ip:$port_web/$endpoint > searchingfor_WP.txt`
+				if [[ $(cat searchingfor_WP.txt | grep Moved) ]]; then
+					result_curl=`curl -s http://$ip:$port_web/$endpoint/ > searchingfor_WP.txt`
+				fi
+				if [[ $(cat searchingfor_WP.txt | grep wordpress) ]]; then
+					endpoint_wp=${endpoint}
+					box_out "We detected a potential WordPress page"
+					echo "[!] Endpoint Name : " $endpoint_wp
+					read -p "[+] Scan WordPress? [y/N] : " scan_wp
+					if [[ $scan_wp == "yes" || $scan_wp == "y" || $scan_wp == "Y" || $scan_wp == "YES" ]]; then
+						wpscan --no-update -o WP_SCAN.json -f json -e vp,vt,cb,u --url "http://$ip:$port/$endpoint_wp"
+					fi
+					break
+				fi
+			done
+		fi
+	fi
+	#IF we have a Smamba server launch a enumeration for users and shares with enum4linux
+	if [[ "$samba" = true ]]; then
+		echo ""
+		box_out "SAMBA SHARE DETECTED"
+		read -p "[+] Scan Samba Server? [y/N] : " scan_samba
+		if [[ $scan_samba == "yes" || $scan_samba == "y" || $scan_samba == "Y" || $scan_samba == "YES" ]]; then
+			command enum4linux -U -S "$ip"
+		fi
+	fi
+
+	if [[ "$ftp_anon" = true ]]; then
+		echo ""
+		box_out "Anonymous FTP login allowed 
+		(For username use: Anonymous)"
+		read -p "[+] Login now? [y/N] : " login_ftp
+		if [[ $login_ftp == "yes" || $login_ftp == "y" || $login_ftp == "Y" || $login_ftp == "YES" ]]; then
+			command ftp "$ip"
+		fi
 	fi
 fi
-
-#IF we have a Smamba server launch a enumeration for users and shares with enum4linux
-if [[ "$samba" = true ]]; then
-	echo ""
-	box_out "SAMBA SHARE DETECTED"
-	read -p "[+] Scan Samba Server? [y/N] : " scan_samba
-	if [[ $scan_samba == "yes" || $scan_samba == "y" || $scan_samba == "Y" || $scan_samba == "YES" ]]; then
-		command enum4linux -U -S "$ip"
-	fi
-fi
-
-if [[ "$ftp_anon" = true ]]; then
-	echo ""
-	box_out "Anonymous FTP login allowed 
-	(For username use: Anonymous)"
-	read -p "[+] Login now? [y/N] : " login_ftp
-	if [[ $login_ftp == "yes" || $login_ftp == "y" || $login_ftp == "Y" || $login_ftp == "YES" ]]; then
-		command ftp "$ip"
-	fi
-fi
-
 #Echo all the results
 banner
 box_out "Final output: "
@@ -332,6 +338,4 @@ echo -e "[+] Samba Server found? : " $(check_result $samba)
 echo -e "[+] Hugo post created?  : " $hugo_created
 
 
-rm endpoints.txt output_nmap searchingfor_WP.txt 2> /dev/null
-
-exit 0
+rm -f endpoints.txt output_nmap searchingfor_WP.txt 2> /dev/null
